docker-app/web-service/requirements.txt
Flask==1.0.2
cassandra-driver


